Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

It looks good to me.  Seems to match the style of gingerbread better.  Is it not possible to drop this into framework somewhere to utilize it more globally for other things as well.

Come to think of it, why don't browser and deskclock share the same resources for these drawables?

Patch-set: 1