Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

works for me, much better than trying to scramble to find the button on the screen when half asleep! 

settings are a bit hard to find, i was expecting them in the settings app not deskclock settings - which you can only get to when setting an alarm, which seems odd in itself.

Patch-set: 3