Update patch set 3
Patch Set 3:

Simon, 
In the settings app there's a section for date and time, this setting doesn't belong there. I feel adding a alarm settings section in the settings app just for this doesn't make much sense either, so the deskclock app settings seem to be the best place.

Patch-set: 3