Fix 5734: Set alarm clock to switch profile
This change creates a new setting for the alarm, that allows to associate
a profile with an alarm, and change to this profile when the alarm triggers.

Patch 2: Also add null profile in HandleSetAlarm

NOTE: The database of alarms is upgrading from version 5 to version 6, so all
alarms are resetting with this change.

Change-Id: Iaf2a24e03d404dc69a90839974a61096f85ae663