Fix 5922: Alarm not being set my Voice Action (from ics)
ics: http://review.cyanogenmod.com/#/c/20463/1

When an alarm is create with the HandleSetAlarm activity (for example with
Voice Search) alarms is created in the database with the enabled flag active
but no the alarm is not enabled.

This changes enables the alarm by default (if no other flag is passed to the activity).

Change-Id: Idde8a227f6c288f40a0a30891148fe9c5d762865