Update patch set 1
Patch Set 1:

David, the main purpose of DeskClock application, and concretely the Alarm activity is
to be be used like an alarm. IMHO, reduce his main functionality is out of his purpose.

Maybe, is a better approach to create an activity for planning task, like a front-end
from crond or atd. Are you interesting in that? Maybe, we can work on that if people 
are interested.

Patch-set: 1