Update patch set 2
Patch Set 2:

Ok. I will play with Notification.BigTextStyle, but i think that implement "show current countdown progress" will required the use of RemoteViews, because the TimeClockService only provides system time diff (to avoid to use the cpu while not user not in the UI) and a PendingIntent to notify the end of countdown. I think following RemoteViews pattern, a Timer view can be attached to it and refresh the notification without asking the service for the current time (and without recreating the notification), and on start/stop of timer will recreate the notification and his view. I will try to join both.

Björn, i not a good graphic designer, so please, if you could submit an icon that matches with the design guidelines, i will be grateful.

Patch-set: 2