Update patch set 3
Patch Set 3:

Just my opinions as a user looking at this screenshot:
https://www.dropbox.com/sh/aikob4s13louit7/Zqx55nvbgJ/device-2012-08-25-120605.png
The background image scrolling in a repeated fashion bothered me because the glaring discontinuity of the background doesn't look good. Also, there doesn't seem to be tabs on top or anything to indicate to a new user that a swipe would do anything, nor is there anything to even indicate that the timer and countdown functions exist.

Patch-set: 3