Update patch set 3
Patch Set 3:

Jorge, I agree that tabs should not be implemented. I was just suggesting that perhaps something should be there to indicate that there is more than just a clock and alarm. Maybe some kind of transparent arrows on both sides that only show when the user touches the screen? Just throwing it out what comes to mind.

Patch-set: 3