Update patch set 5
Patch Set 5:

Working on a Swipey Tabs like indicator. But this need some time.
The idea is have a transparent tab (text + indicator bar) that holds visible some seconds and then disappear. When user starts swiping between
views the tab is show again and move together with the fragment.


Björn, i don't understand your comment. The application has 2 icons: 
- The first is showing in the title bar of the fragment (Markus suggest remove it when viewpager indicator is developed) and in the Notification as the large icon.
- The second one is only showing in the statusbar, for the Notification small icon. I opted by give an gray color aspect like deskclok alarm icon.

I attach an screenshots of the notification when the countdown is running. The images shows the 2 icons. Let me know what of the icons is incorrect.

https://www.dropbox.com/sh/uujna6c4yb2lftr/yXBQxKBAD2

Patch-set: 5