Update patch set 7
Patch Set 7:

heh, I actually started on a project like this, once.
My attack vector was an Activity which contained an ActionBar, and then make AlarmClock Stopwatch, Countdown, Worldwatch etc. inherit that first activity. iow. staying out of the way of the fullscreen desckclock activity.

I started somewhat to migrate everything to Fragments as well, but well.. It would create a merge hell when 4.1+/5.0 comes.

Patch-set: 7