Update patch set 7
Patch Set 7:

David, do you have any suggestion? I haven't any idea on how to minimize the transition effect from a wallpaper-fullscreen app to a non wallpaper-nofullscreen app. I think that if holo background is used, a title bar or something that identifies the screen is needed. Maybe, the problem is the holo background, but with a wallpaper, the timers fragment doesn't look good.

Patch-set: 7