Update patch set 7
Patch Set 7:

@Jorge,

I'm not sure what the best solution would be.  Suggest we keep it as is (just fix the alignment) and see.

That being said, I would suggest placing both the new pages on the right of the deskclock (i.e. deskclock is the first page).  That way, once the user has swiped to the the 'holo' pages, he/she doesn't have to swipe back over the deskclock to get to the other page. 

I really like the tutorial you added :) Would you be willing to create similar tutorials for some of the other features we are adding/have added to CM10?

Patch-set: 7