Update patch set 8
Patch Set 8:

Looks great on maguro now, no issue with the icon at the top being obscured by the status bar anymore.  My only request is that the two extra pages you added both be to the right of the clock, rather than on either side, but that is a minor gripe.  I love the addition of this functionality to the stock clock app, its been sorely missing.

Patch-set: 8