Update patch set 9
Patch Set 9: I would prefer that you didn't submit this

Please move the Stopwatch to the right of the main screen so that the user swipes right for both the stopwatch and the countdown.  The main clock/alarm screen should be the first screen.

I would also suggest a faster time-out on the fade of the page indicator.  It is too slow.  Should match the speed of the home screen indicator fadeout.

Patch-set: 9