Update patch set 12
Patch Set 12:

As far as the ui on the timer, I think including a long press at all is unnecessary.  The number fields should each just be scrolling number pickers that lock once you start the timer, otherwise they remain adjustable.  Arrows above and below each field or a partial view of the next/previous number would be enough of a hint that the numbers are adjustable and is much more discoverable than a long press.

Patch-set: 12