Update patch set 14
Patch Set 14:

When trying to add a new Countdown timer the value of the timer is only saved when touching the plus symbol.

When touching the "Add Current Timer" text the countdown value is reset.

That's not the behaviour i expected. I think touching the text should add a new timer as well. Or there should be no action at all when touching the text.

Patch-set: 14