Update patch set 15
Patch Set 15:

I think being able to set milliseconds should be removed from timer. I don't really see anyone using it, and resetting it is annoying if you want to change it to 0 after stopping it.

IMO it kind of makes the UI a little crowded when you are setting the timer.

Patch-set: 15