Update patch set 3
Patch Set 3:

Kim, i think that you describe, it's not a bug. The behaviour should be:

- All the app have a screen wakelock (the screen never should be turn off). This comes from the old AOSP DeskClock activity.
- DeskClock fragment. Has 2 differents behaviours:
     * When the user press the screen, the screen should be dimmed/undimmed
     * After 5 minutes of user inactivity, the clock should be enter in screenmode
- Stopwatch & Countdown fragments. After 1 minute of inactivity the screen should be dimmed

I have no problems with this behaviour. In fact, the last change could cause the opposite, that the screen turns off because the wakelock code will not be reach.

If you could reproduce the crash, and you can catch a log, please post or send it to me, and i'll try to see what it's wrong.

Patch-set: 3