Clock: fix layout for 320dp devices
800x480px dpi devices have less screen space than the xhdpi devices
Google designed the layout for and the layout needs adjustment.

The number in the timer (left-most tab) is cut off on older devices.
Reduce the font-size so that it fits properly.

This should not have any effect on devices with larger screens.

(cherry picked from commit 1bb1f7e49509d88d6cae464384fa93b8f6819ee9)

Change-Id: Icc9defddd119bc95fbd70c1eebe906353c2c2f91