Update patch set 3
Patch Set 3:

IMHO the default and expected behavior should be:

Stopwatch: you are watching the screen to stop the watch as soon as the external event you are observing hits it's end or milestone you want --> screen needs to be always on.

Timer: Times logic is the exact other hand of stopwatch. You know how much time you need before interfering into the external event you are observing. You can do this by looking into the display directly or being alarmed as soon as it reaches the pre-defined amount of time that you set --> screen doesn't need to be always on.

This is what I think the logic imposes and the user would expect, but as we already have a setting to rule it im not gonna be against it.

Just my two cents :)

Patch-set: 3