Update patch set 3
Patch Set 3:

Yep, that is it. 

Since its the expected behavior as I exposed, I even think that we can make it default behavior without living it optional. This would be good if our intention is to make a more "ready out of the box" product. But sure option won't hurt either. Maybe David or Danny can rule about option or not. :)

Patch-set: 3