Update patch set 4
Patch Set 4:

Pedro:  Screen turn off bug: I will check the source code of the DeskClock next time. I don't promise to spend too much time with it, I think it is a "whocares" category bug, and there is a new option to keep the screen on. :)

Action button for settings: It is a good idea.

Thank you

Patch-set: 4