Update patch set 1
Patch Set 1:

I like the general idea but it's not really centered. It's just centered is the space that is left from screen width - menu botton width. I think it would look much better wehn it is centered from the hole screen width.

To do that the start button needs a pedding which is the half of the menu button. 32 dip when I'am not wrong.

@David: What do you think about that?

Patch-set: 1