Update patch set 1
Patch Set 1:

I tried adding some padding to "really" center the text, but personally I that looks really strange when you press the button and the text is off-center in the highlighted background.

That could be solved by doing some relative layout trickery to float the menu button above the cancel & start buttons, but that could be problematic on small screens where they could overlap completely.

Suggestions are welcome. I think that the current patch set looks better than the current layout and is a fairly reasonable solution.

Patch-set: 1