Revert "Make the settings menu accessible on every tab"
This commit has unintended consequences with the UI layout of the start/stop button of the timer (see http://i.imgur.com/G4eNx2Z.png )  The button used to be centered and is now off center next to the menu button.

I am not sure what the use case is for this commit (i.e. what user problem we are trying to solve here) but IMHO the end result is not worth it.

This reverts commit 057d88c735d8d0e7dce9714f9db28388501c24e7

Change-Id: I2ffbba4cf6f46ed06d67b848e482b484c1ca4c92