Update patch set 1
Patch Set 1:

@Veeti, my first question would be "what problem are we trying to solve and how will this make things better for the end user".  If there are no clear answers, to the above we should consider not doing it.

My second question is whether it is something that would be a unique feature of CyanogenMod (e.g. Expanded desktop or Quick reply) or needs to be part of AOSP (drawable, layout, general UI).

If it passes the first question and would primarily benefit AOSP, we should submit it to AOSP first and see what Google does with it.  If they accept it, we include it, if not, we typically don't since we don't want to complicate the future merge of Android 4.2.2 or 5.0.

Patch-set: 1