Update patch set 1
Patch Set 1:

Let me illustrate this problem with a real-world scenario that just occurred to me. I often like to take "power naps" using the timer feature in the Clock application, and thus a couple of hours ago I went to type 30 minutes into the timer and pressed "START".


However, something had changed. Usually, the start button covers the entire screen width. Now it's been crammed to the left corner and the touch area only covers half of the screen width.


I missed the button completely by clicking the large, gaping area to the right assuming that it is still a touchable part of the button (and that's where muscle memory also tells me to click). Fortunately I noticed this before falling asleep and corrected myself. (I did inevitably end up oversleeping a couple of hours anyway, but like many others that's an issue in my head and not CyanogenMod).


Not only is this very detrimental to the usability and typical user expectations of the timer screen, it also looks pretty terrible as it is now.


This does not apply to AOSP because as far as I can gather this issue was created by an earlier patch that adds the menu/action overflow/whatever-it's-called-nowadays button to each one of the three panels. (AOSP only shows it on the center panel). As for merging, this brings the layout _closer_ to AOSP than it is now by removing the spacer element.

Patch-set: 1