Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I have submitted a revert for the original commit that added the menu button to the additional screens (see http://review.cyanogenmod.org/#/c/31783/ )  I am -1'ing this commit as a result.

Thank you for doing this but we should not have caused the problem in the first place by merging the other commit.

Patch-set: 1