Update patch set 1
Patch Set 1:

@David, the Start/Stop button is supposed to be centered at the bottom of the stopwatch.  With the menu button enabled (which I also question what real user problem enabling it solves), the button is now misaligned and looks unbalanced and out of place.   Same goes for the timer screen.  

As I said above, I question why we would want to add the menu button to these screens in the first place - the main screen that shows when you enter the app has the menu button nicely aligned at the bottom with the other 2 buttons.  It makes sense where it is and looks balanced.

The DeskClock design is one that was used by the Android Design team (and Roman Nurik) to showcase a proper design of a Holo app.  We are breaking their design (and the showcase app) by doing this.

Patch-set: 1