Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

@David: It started from here: http://review.cyanogenmod.org/#/c/30664/
please check the comments (Jan 24-26)

The DeskClock application has 3-4 screen and IMHO it is not an intuitive application. It contains more different ways to navigate between the screens (e.g. alarm to clock, alarm to settings, clock to settings: you have to use the back button, click to title (alarm), tab, menu, settings button) It is a design question: Why the screen turns off when the stopwatch works (in the original application)? The accessible settings menu try to make this application cleaner for the average user.
I am surpised that this application has a strict design concept. I am happy in this case and I don't want to violate this concept. :)

Patch-set: 1