Update patch set 1
Patch Set 1:

@David, I still ask the question "what is the problem we are trying to solve" when I look at adding menu buttons to the other 2 screens - the original design has the main entry point of the app (clock/alarm) with the menu item, and the menu refers to the app as a whole.

The menu is not applicable to the timer and the stopwatch screens (the settings menu contains clock and alarm settings and only one stopwatch setting) - it does not make sense to show those settings items within a feature that they do not apply to.

IMHO we are not really improving the UX by adding this feature.  In fact, we are breaking the 'balance' of the screens and potentially introducing confusion by showing settings that are not applicable to the screen the user is on.

Patch-set: 1