Update patch set 1
Patch Set 1:

Ok, I see. There are only the "Keep the display on..." settings which is interesting now in the stopwatch tab, but it is not an every day used option. The settings menu on this pages will make more sense later, if we will more options on the settings page which is related with the timer or the stopwatch. On the other hand it would be comfortable when the user can access the settings (and every screen from the DeskClock) everywhere, immediatly. I have started to believe that the alarms screen is the odd one out in this application which is not okay.

Thank you for your explanation.

Patch-set: 1