Update patch set 1
Patch Set 1:

This is maybe grammatically correct :)

There is only the "Keep the display on..." setting which is interesting now in the stopwatch tab, but it is not an option used every day. The settings menu on these pages will make more sense later, if we will have more options on the settings page which are related to the timer or the stopwatch. On the other hand it would be comfortable when the user can access the settings (and every screen from the DeskClock) everywhere, immediately. I have started to believe that the alarms screen is the odd one out in this application which is not okay.

Patch-set: 1