Update patch set 1
Patch Set 1:

I don't like the two in one line solution. I like this a lot more. I also improves the freedome of translations in some langueges. In some languages (e.g. German) it's quit difficult to find a translation which descibes the feature well enough and is short enough to fit in the line.

Patch-set: 1