Update patch set 4
Patch Set 4:

I like this too, however I am not sure it should be merged into CM. 

Sleep cycle timeframes are not a fixed 90 minutes.

According to Wikipedia, most humans reaches REM sleep within 90 minutes, but the sleep cycles themselves vary from 90-120 minutes. 

Hence, without the proper scientific weight, this app would be little more than a gimmick.

That said, I havent really tried a lot of the sleep cycle apps on the Market, so maybe they are all so terrible that this is a better alternative anyways. Just not sure if it should be forced upon us into a ROM.

Great work though and its a cool idea!

Patch-set: 4