Update patch set 4
Patch Set 4:

Well yes, but the estimate would still, from what I understand, be so inaccurate that it doesnt really provide any additional usability. 

Wikipedia: "In humans, the average length of the first sleep cycle is approximately 90 minutes and 100 to 120 minutes from the second to the fourth cycle, which is usually the last one"

So if you sleep for 8 hours, the app would calculate 90 minutes for each cycle, meaning the time estimates it would provide could be completely off, hence my concerns. 

But Im just voicing my opinion. Its up to the higher powers to decide ;)

Patch-set: 4