Update patch set 4
Patch Set 4:

Forced upon the ROM? Ain't this an option? There is nothing forced. 

The 90 min cycle is an average that is totally acceptable. This feature is intended to - try to help - you have a better sleep. I really can't see no polemic about it.

Also, its not a question of replacing bad apps on play store. Its just CM providing an additional feature and one less app on the drawer for those who use third party.

Patch-set: 4