Update patch set 4
Patch Set 4:

I disagree that a fixed 90 minutes is usable, when most have a sleep cycle that would vary between 90 and up to 120 minutes during a lengthy night sleep. But again, this is not my decision to make, and if it gets merged I am free to not use the feature of course. I just want to point out that scientifically, from how I understand the subject of sleep cycles, this would be little more than a gimmick. 

Side note: Another issue is that timing when you actually fall asleep is quite difficult and hence I think it makes more sense to have an app where you tell it approximately when you want to wake up, and based on the microphone and other sensor data, the app would estimate when you fall asleep and when you reach REM and then wake you up within a certain interval around the specified wakeup time.

Patch-set: 4