Update patch set 4
Patch Set 4:

You can't make the interval user customizable, because again, the sleep cycle varies for everyone all through the night. Meaning the first cycle might be 90 minutes, the next one 110 minutes, etc.. 

And I think its important to be strict as to what custom features to add and not. There are other ROMs out there with more experimental features if that's your cup of tea.

Patch-set: 4