Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

Wikipedia is not the ultimate trustable source to call out the "scientifically" argument.

Even if the info is 100% correct, we could use a 105min cicle to feat the average between 90 and 120 and reduce the chance of error since the important is to be to most close to cycle end as possible. A configurable cycle would be good here.

Aside of that are making a so big deal for a simple incremental feature that does exactly what it describes it does: Try to help on waking up based on a 90min cycles.

The feature is great.

Patch-set: 4