Update patch set 4
Patch Set 4:

I attended a couple of  classes on sleep theory, and while it wasnt anything too complex, it did support what the research referred to on Wikipedia states. The sleep cycle time is variable and varies from person to person, so I think its simply not possible to predict it in an accurate matter. 

I dont want to make a big deal out of anything, and I wont complain if this feature gets merged. Im just pointing out that if the point of the feature is to help you wake up when you're in N1, then it simply wont work reliably at all, based on what little I know about sleep theory that is. Based on that rationale, the interval used might as well be grabben from /dev/random, as the accuracy would be pretty much the same. If anyone knows more about this than I do, please add your input. 

It just seems like a gimmick to me..

Just my two cents..

Patch-set: 4