Update patch set 4
Patch Set 4:

Statistically, if you pick any number between 90 and 120 you will be more accurate than picking a random number out of that interval...

Human body doesn't absorve drugs on the same pace on every person and labs doesnt make person-specific drugs for this. They use averages.

The human v02 max, % of fat in the body and a gazillion other things in the world are calculated based on assumptions and estimation when intervals between samples are sometimes way bigger than 90-120 range. As I told, if you want better accuracy we can use 105 and call it a day.

To do a simple exercise, I generated 5 series (like 5 cycles) of 1000 random numbers between 90 and 120 with average on 105 and the std dev on 15 (120-15=105) assuming its a bell shape distribution. The volatility between the cycles leght on this simulation was 13%. Seems well controlled to me.

Patch-set: 4