Update patch set 5
Patch Set 5:

This function recommends times for when do you go to the bed. It works how the commit comment describes, so this function works as expected. 

The optimal bedtime depens on 1000000 things, so (I think) that is a bad conslusion that this function has to provide exact bedtime values with a minute accuracy (and maybe with a dinner recipe, because it is important to sleep cycles). 

Personally I like this function, and I think there are more user who likes it than not. That is why the feature could be valuable. This is only my opinion.  (... and of course the recommended bedtime value works for me +1 :))

Patch-set: 5