Update patch set 5
Patch Set 5:

And to take what was mentioned here earlier: "The optimal bedtime depens on 1000000 things" - Soo, lets not merge a feature that tries (and fails) to do just that. If the purpose is to help you wake up more refreshed, then a better approach would be to wake you up within a given timeframe.

Patch-set: 5