Update patch set 5
Patch Set 5:

@Edvard: "A better approach would be to dictate when you should wake up based on your level of consciousness (movement, etc)."

This is not only about this patchset: We will never ever have better solution if abandon the first changeset. Potentially this patchset is the first version of the better solution what you suggested.

This feature is very simple but likeable. This has a good place in the DeskClock, using it is not mandatory. We can revert/abandon it later. 

(+ I understand everything what you wrote but I'm not sure that this patchset is completely wrong).

It is only an opinion.

Patch-set: 5