Update patch set 5
Patch Set 5:

Agreed that this is placebo. Alarm has no way of knowing when you actually went to sleep.

Other alarms that do this use a pre-alarm that isn't nearly as loud to attempt to wake up the user if they're not in a deep sleep.

I think that's mostly hokey as well though.

Patch-set: 5