Update patch set 5
Patch Set 5:

The calculator determines sleep times based on widely accepted averages. 
It isn't meant to be exact; although, I think adding options to adjust 
it according to each user could be a nice solution to that. The idea 
here is to prevent the user from setting an alarm that would wake them 
up during deep sleep. 

I think it's a solid feature for CM. It provides room for expansion and 
reduces the number of third party apps users have to download. In the 
long term, it will improve the clock app. 

It isn't the calculator's responsibility to know when you fall asleep. 
It's the user's to go to bed on time. Also, regardless of the science, 
people use this feature in other apps. It will make CM look better 
whether it actually improves it or not.

Patch-set: 5