Increase DeskClock db version to be compatible with CM10.0
In cm-10.0, the db version for DeskClock was already 7, but in cm-10.1,
it's set back to 6. This causes the DeskClock to break if upgrading from
cm-10.0 to cm-10.1. This fix increases the db version to 8, and drops
the table during the upgrade. Although this will lose all alarms already
setup, the code is simpler than trying to handle all the possible db
version mismatches.

Change-Id: I9a4d849f5145acb798cb8e87f1c35fbf26eaac05