Update patch set 2
Patch Set 2: Looks good to me, approved

The main reason for this is that putting mixing AOSP and CM translations are causing merge conflicts, when an update of AOSP Android is released (such as 4.2.2 for cm-10.1) and the team wants to update cm-10.1 for it.

We currently have a similar situation: right now there's discussion about the opt-in/opt-out of sending anonymous non-personal data to Google Analytics. The strings for this option were merged a while ago, meaning the strings were sort of at the top of the CyanogenMod additions section. Since these strings have been been removed, some people have uploaded new commits (as Settings is still under development). But suddenly the strings were added back, as the whole situation got reverted since some people didn't like the way it went. This meant merge conflicts for all those patches and everyone had to rebase manually.

BTW, I have added you to the Google+ Community for CyanogenMod Translators.

Patch-set: 2