Update patch set 8
Patch Set 8:

IMO, yes. But should be UX whom take the decision.

I want notice you that according to the latest problems with the yahoo api, this changes uses maps.googleapis.com/maps/api/timezone to determine the user timezone, with a limit of 2000 calls per day. Would be a problem?

Patch-set: 8