Added music alarm feature
Added ability to playback a music file instead of an alarm.
I added a Song Picker activity and made some changes to the
alarm clock layout, to switch between alarm and music you
need to click the icon next to the ringtone name textview.

The Song Picker activity returns the file's URI so it works
the same way as normal ringtone, no changes have been made
to the process of setting an alarm and the klaxon.

Change-Id: Idf4b17c9bd55a761a9e6c0a1ed149121a11e49ca