Update patch set 6
Patch Set 6: Code-Review-1

I had a quick look and I see that you modified a lot of lines that really don't need to be changed. If you wish to clean up whitespace and such please do so on a different commit, otherwise reviewing the actual, meaningful changes is a pain.

As an example, your commit changes over 100 lines on strings.xml but it's all whitespace/indentation changes (or if there's any actual change, I just didn't see it because of all the noise)

Patch-set: 6