Update patch set 6
Patch Set 6:

(19 comments)

Hi! I changed almost everything that was asked and I'm currently working on implementing a Loader and CursorAdapter in the AlarmSongPicker Activity.

I also need to remove the strings.xml file since there was no changes to it after all, what could be a good way to do this without deleting the file or breaking the repo? Thanks!

Patch-set: 6