Update patch set 7
Patch Set 7:

That could be cool but it wouldn't work for everyone. I currently have Apollo disabled because I use Google Play Music as my music player, so in my case I couldn't set a song as my alarm.
Maybe if we set an Intent to start a picker activity? (ACTION_PICK_SONG?). So if you have multiple apps it allows you to select which app you'd like to use or, like in my case, it falls back to the AlarmSongPicker activity. But I'm not sure what needs to be done to accomplish this.

Patch-set: 7