Update patch set 7
Patch Set 7:

Yes, I know but I wasn't sure if it was planned to remove the AlarmSongPicker and just use the one from Apollo (which would not work for everyone). So as a solution I proposed to use an Intent, to choose from Apollo (if enabled) or the AlarmSongPicker.

Patch-set: 7