Update patch set 10
Patch Set 10:

Much better. Fetching the album art still happens in the main thread, though. As this is an expensive operation, it should be done in the main thread as well. I need to look up what the best way to do that would be, but I guess it would be an AsyncTask fetching the album art into a ID -> Bitmap map. I guess it should be possible to have a cursor that combines that map and the DB cursor.

Patch-set: 10