Update patch set 1
Patch Set 1:

I changed to shorter words because the activate/inactivate it's too long. This makes that the button cannot be drawn on the screen completely because it is too wide so it's out of the screen. That's why I changed.

Patch-set: 1
Reviewer: Arnau <6733@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0