Update patch set 1
Patch Set 1:

I have no screenshots. The :00 and :30 Buttons are on the alarm time chooser, from where I stealed the idea shamelessly (I wondered they were missing in the timers). I left out : as they can also be uses in the middle of the timer time (while in the alarms, they only can be used for the minutes).

Patch-set: 1