Update patch set 6
Patch Set 6: Code-Review+1

Pavel, I'm ok with this changes, but as a suggestion don't push translation code into a featured code. IHMO, it is better, for rebasing or cherry-picking purposes for next release, do not depend in translation strings that can cause conflits. Also, putting your translation in a separate commit allow russian translation to discuss about the strings.

Patch-set: 6