Fix infinite loop with missed alarms
Looks like the snooze code got copied for missed alarm instead of the
more appropriate dismissed alarm. updateNextAlarm sends out a broadcast
of the next alarm to be fired, but unfortunately if there are no other
alarms set it sends out a new broadcast of the same missed alarm,
resulting in an infinite loop.
Copy the dismissed alarm code instead to delete the alarm after the
missed notification is sent to the status bar.

Change-Id: Iddb784c8745a50b9c95d15d60b24c4365b666b54