Update patch set 1
Patch Set 1:

The way the code is supposed to work is it sets the alarm state to ALARM_MISSED, sets up the alarm missed notification, which changes the state to ALARM_DISMISSED, and then sends out the updateNextAlarm broadcast.

For some reason though, the notification code is not dismissing the alarm before it returns here, leading to the infinite loop. 

Honestly, I don't see why we can't just dismiss the alarm here instead.

Patch-set: 1