Update patch set 1
Patch Set 1:

I think someone mentioned somewhere else (IRC maybe) that the problem is we are setting 2 different alarm types now, ALARM_RTC_WAKEUP and ALARM_RTC_POWEROFF_WAKEUP so the current code which schedules the alarm to be cleared in the future doesn't work because there are two alarms set. (whereas my code change which kills it immediately seems to work). I wonder if detecting if the host supports POWEROFF_WAKEUP and only setting one or the other but not both would solve this

Patch-set: 1