Update patch set 1
Patch Set 1:

@Jorge The RTC alarm is oneshot, so it's only going to fire once. So the impact of the infinite loop there is minimal. 

I'm not saying my fix isn't a hack. I'm saying the current code is dumb regardless of what is causing the loop. Let me explain.

So an alarm is missed, either because the time for it is passed (AlarmStateManager line 579) or because the alarm instance has a state of MISSED_STATE (line 667)

Both of these call setMissedState(). Let's look what happens there.
The alarm is stopped if it's going off.
The instance parent is updated (so say it's a recurring alarm, you need to schedule the next instance)
The alarm state is changed to MISSED_STATE. Remember, the *only* code that checks for MISSED_STATE calls this very function. And the only code that sets it is this function. Dumb already setting in.
Next we call AlarmNotifications.showMissedNotification to notify the user of the missed alarm. 
It sets up the notification and does a very curious thing. I don't pretend to understand this part, but it looks like it it deletes the alarm intent if it's set to DISMISSED. 

then we return to setMissedState where the alarm state is set to DISMISSED (presumably for setDismissedState) to take care of, and then updateNextAlarm is called, which is supposed to take the next valid alarm and runs with it.

Instead what appears to be happening is setDismissedState isn't getting called fast enough and updateNextAlarm notices the SAME missing alarm and broadcasts it, causing setMissedState to get called again, causing an infinite loop and royally chewing through battery.

So let's look at the levels of dumb. First we set the alarm state to something that CAN ONLY TRIGGER THIS SAME FUNCTION TO BE CALLED. Then we set it to dismissed state and hope setDismissedState gets rid of it before it fires again. My patch completely eliminates the dumb. We already know we have a missed alarm, and we sure as hell don't want to call setDismissedState again. SO WHY SET THE STATE TO MISSED?

and why the hell set the state to dismissed and hope the dismiss function picks up quickly enough? Does that make any sense? Isn't it more logical to either directly call setDismissedState or simply kill the alarm in the exact same manner right here? There is absolutely no benefit to the system to schedule a pending intent change on an alarm we already know is in the past and doesn't need to be fired again.

That's all my patch does. It sets up the user notification and kills the alarm, instead of setting up a pending intent and hoping it gets dismissed before it fires again.

Patch-set: 1