Update patch set 1
Patch Set 1: -Code-Review

The problem is that the new pending intent doesn't pass a valid stateAlarm AlarmStateManager#handleIntent, calling registerInstance instead of setAlarmState, causing and infinite loop, since the alarm is fired until is dismissed.
This is a waste of resources while the alarm is in a fired state.

I'm going to remove my -1, but i think this should be fixed properly, because not all devices has supports the new pending intent. I think is better to revert the previous change.

Patch-set: 1