Update patch set 9
Patch Set 9: Code-Review-1

(2 comments)

Look goods to me. The -1 is mainly for the close cursor comment.

Kiril, do we need a Multiplayer for this patch? IIRC the old cm-10.2 patch could handle a playlist, but this one only handle a random song, so couldn't we set a random uri in alarm database and generate a the randim media uri when MediaPlayer is going play the song? Just a suggestion.

Patch-set: 9