DeskClock: Search cities with any part of the name
Only the the starting part of the name could be used to search.
Therefore, when try to search the cities with the middle or ending
part of the name, no expected result will be listed.

Use contains() method instead of startsWith() to match the search,
so that all the names containing the search string will be listed
in the result.

CRs-Fixed: 637280

Change-Id: I670a163b9c4df575445b58eeb1cc5070a557550e