Update patch set 1
Patch Set 1:

This statement was introduced in the change: 0a17d0168720f00fd19e5ab267b4a26b37260bf8 "deskclock: take persistent grants of external uris"

which appears to be cherry picked from this cm11 change:

http://review.cyanogenmod.org/#/c/72132/

Since that code is not in the original change I'm guessing this was a merge error.  If anyone knows otherwise let me know.

Patch-set: 1