DeskClock: Handle invalid default alarm URI
If the user sets an external file as their alarm and the file is
subsequently deleted, DeskClock will go into a crash loop trying
to reset to a default alarm uri (which was set to this external
file).

This patch will make so that when this scenario is triggered, and
the default uri is invalid, DeskClock will fetch and reset default
alarm URI from the system property 'ro.config.alarm_alert'. If
no such system property exists or the file cannot be found, it will
pick a random alarm tone from MediaStore.

Change-Id: I481876e0a25fbbb7ab1be8e3d7c9f8a94d66a43c