Update patch set 1
Patch Set 1:

As a temporary solution this can be merged, but I need someone to write me a script that converts string arrays to string references.

As discussed in the past, we are going to create a new untranslatable string array with all entries converted to string references, so we only have to maintain only this array instead of all arrays in all languages when we add a new entry.

The script I mentioned is needed so we can keep the existing AOSP translations

Patch-set: 1