Update patch set 3
Patch Set 3:

I see your point, but I would like to avoid multiple sensor checks.  So I believe what you're saying is that if I set this on a device that doesn't work, then get this patch it will be set?  Maybe in the same logic where it is hiding the view, it can just force overwrite the value.  Thoughts?

Patch-set: 3