Update patch set 3
Patch Set 3:

Okay, think I know what's going on. Looks like when you're building the patch, you pushed and force-dex-opted. But you didn't restart, so the manifest changes did not get picked up, and the receiver does not get registered. I'm not sure why exactly, but this seems to mess up global ids for the intent - so the AlarmStateManager doesn't respect the pendingIntent fired from the notification .

Just tested this with a QRDevice and seems to be working! :)

Patch-set: 3