Update patch set 1
Patch Set 1:

Having the power-off-alarm to power on the device 120 seconds before actual alarm time is nice to it hit the time more correctly, but there's no guarantee the device will boot in 120 seconds or any seconds. e.g. after an OTA update it may reboot but the user decides to pull the battery, he didn't know it needs a more than 120s upgrade that will miss the alarm

I strongly recommend to handle a missed alarm as firing alarm

Patch-set: 1